% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_ids.R
\name{convert_gene_ids}
\alias{convert_gene_ids}
\title{Convert gene IDs to symbols}
\usage{
convert_gene_ids(results, species)
}
\arguments{
\item{results}{Tidy DE results table}

\item{species}{Species name (eg. "human", "mouse")}
}
\value{
Results table with gene symbols
}
\description{
Convert gene IDs to symbols
}
\examples{
results = dream_to_dge(model) 
results = convert_gene_ids(results, species="human")

}
