% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_to_apeglm_dge.R
\name{deseq2_to_apeglm_dge}
\alias{deseq2_to_apeglm_dge}
\title{Extract shrunken DESeq2 results (apeglm)}
\usage{
deseq2_to_apeglm_dge(
  dds,
  var_interest,
  numerator = NULL,
  denominator = NULL,
  use_adaptiveshrink = TRUE
)
}
\arguments{
\item{dds}{DESeqDataSet after differential expression analysis}

\item{var_interest}{Variable of interest}

\item{numerator}{Numerator value for variable}

\item{denominator}{Denominator value for variable}

\item{use_adaptiveshrink}{Logical, use adaptive prior (default TRUE)}
}
\value{
Shrunken DESeq2 results as tidy table
}
\description{
Extract shrunken DESeq2 results (apeglm)
}
\examples{
dds <- DESeq(dds)
results <- deseq2_to_apeglm_dge(dds, var_interest="condition",
                                numerator=1, denominator=0)

}
