% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_to_ashr_dge.R
\name{deseq2_to_ashr_dge}
\alias{deseq2_to_ashr_dge}
\title{Extract shrunken DESeq2 results (ashr)}
\usage{
deseq2_to_ashr_dge(dds, var_interest, numerator, denominator)
}
\arguments{
\item{dds}{DESeqDataSet after differential expression analysis}

\item{var_interest}{Variable of interest}

\item{numerator}{Numerator value for variable}

\item{denominator}{Denominator value for variable}
}
\value{
Shrunken DESeq2 results as tidy table
}
\description{
Extract shrunken DESeq2 results (ashr)
}
\examples{
dds <- DESeq(dds)  
results <- deseq2_to_ashr_dge(dds, var_interest="condition", 
                              numerator=1, denominator=0)

}
