% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_to_dge.R
\name{deseq2_to_dge}
\alias{deseq2_to_dge}
\title{Convert DESeq2 results to tidy table}
\usage{
deseq2_to_dge(dds, contrast)
}
\arguments{
\item{dds}{DESeq2 object after DE analysis}

\item{contrast}{Contrast coefficients}
}
\value{
Tidy tibble with shrunken DE results
}
\description{
Convert DESeq2 results to tidy table
}
\examples{
library(DESeq2)
metadata <- tibble(sample = c("s1","s2","s3"),
                   type = c("treated","untreated","treated"))
counts <- tibble(gene_id = c("ENSG1","ENSG2"),
                 s1 = c(10,20), 
                 s2 = c(15,50),
                 s3 = c(20,100))
dds <- DESeqDataSetFromMatrix(counts[,2:3], metadata, ~ type) 
dds <- DESeq(dds)
results <- deseq2_to_dge(dds, contrast=c(-1,1))  

}
