% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{dge_to_gsea}
\alias{dge_to_gsea}
\title{Perform gene set enrichment analysis}
\usage{
dge_to_gsea(
  dge.res,
  nPermSimple = 10000,
  sig.list = NULL,
  species = "Homo sapiens",
  msigdb_category = "H",
  msigdb_subcategory = NULL,
  minSize = 30,
  save_output = NULL,
  var_interest = NULL,
  numerator = NULL,
  denominator = NULL,
  output_filename = NULL,
  ...
)
}
\arguments{
\item{dge.res}{Tidy DE results table with fold changes}

\item{nPermSimple}{Number of permutations for p-value calc (default 10000)}

\item{sig.list}{Optional named list of gene sets to use (default NULL)}

\item{species}{Species for msigdb sets (default 'Homo sapiens')}

\item{msigdb_category}{msigdb category to use (default 'H')}

\item{msigdb_subcategory}{msigdb subcategory (default NULL)}

\item{minSize}{Minimum gene set size (default 30)}

\item{save_output}{Logical to save results (default NULL)}

\item{var_interest}{Variable of interest if saving file (default NULL)}

\item{numerator}{Numerator value (default NULL)}

\item{denominator}{Denominator value (default NULL)}

\item{output_filename}{Custom filename for results (default NULL)}

\item{...}{Additional parameters for fgseaMultilevel()}
}
\value{
DataFrame with GSEA results
}
\description{
Run GSEA on differential expression results using msigdb gene sets
}
\examples{
results <- deseq2_to_dge(dds)
gsea_res <- dge_to_gsea(results)

}
