% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_to_dge.R
\name{dream_to_dge}
\alias{dream_to_dge}
\title{Convert Dream results to tidy table}
\usage{
dream_to_dge(dream_model, coef)
}
\arguments{
\item{dream_model}{Fitted Dream model object}

\item{contrast}{Contrast coefficients}
}
\value{
Tidy tibble with DE results
}
\description{
Convert Dream results to tidy table
}
