% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_to_hgnc.R
\name{ensembl_to_hgnc}
\alias{ensembl_to_hgnc}
\title{Convert Ensembl IDs to HGNC symbols}
\usage{
ensembl_to_hgnc(results)
}
\arguments{
\item{results}{Table with Ensembl gene IDs}
}
\value{
Table with HGNC symbols
}
\description{
Convert Ensembl IDs to HGNC symbols
}
\examples{
results <- deseq2_to_dge(dds)
results <- ensembl_to_hgnc(results)

}
