% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_to_dge.R
\name{limma_to_dge}
\alias{limma_to_dge}
\title{Extract differential expression results from limma model}
\usage{
limma_to_dge(
  fit,
  var_interest,
  numerator = NULL,
  denominator = NULL,
  save_output = NULL,
  output_filename = NULL
)
}
\arguments{
\item{fit}{Limma model fit object}

\item{var_interest}{Variable of interest}

\item{numerator}{Numerator level}

\item{denominator}{Denominator level}

\item{save_output}{Logical to save output}

\item{output_filename}{Custom filename for output}
}
\value{
Tibble with DE results
}
\description{
Extract differential expression results from limma model
}
\examples{
fit <- rnaseq_to_limma(metadata, counts, design)
results <- limma_to_dge(fit, "condition", 1, 0)

}
