% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dge_results.R
\name{merge_dge_results}
\alias{merge_dge_results}
\title{Merge multiple DGE results tables}
\usage{
merge_dge_results(..., suffixes = seq_along(...))
}
\arguments{
\item{...}{DGE results tables}

\item{suffixes}{Optional suffixes}
}
\value{
Merged DGE results
}
\description{
Merge multiple DGE results tables
}
\examples{
table1 <- deseq2_to_dge(dds1)
table2 <- deseq2_to_dge(dds2) 
merged <- merge_dge_results(table1, table2)

}
