% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gsea_results.R
\name{merge_gsea_results}
\alias{merge_gsea_results}
\title{Merge multiple GSEA result tables}
\usage{
merge_gsea_results(..., gsea_result_names = NULL)
}
\arguments{
\item{...}{List of GSEA result data frames}

\item{gsea_result_names}{Custom names for each GSEA result}
}
\value{
Merged data frame of all GSEA results
}
\description{
Bind rows of multiple GSEA result tables into one table
}
\examples{
results1 <- dge_to_gsea(results1)
results2 <- dge_to_gsea(results2)
merged <- merge_gsea_results(results1, results2, 
                             gsea_result_names = c("Treatment1", "Treatment2"))

}
