% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results_long.R
\name{merge_results_long}
\alias{merge_results_long}
\title{Merge multiple result tables into long format}
\usage{
merge_results_long(..., result_names = NULL)
}
\arguments{
\item{...}{List of result data frames/tibbles}

\item{result_names}{Custom names for each result object}
}
\value{
Merged long-format data frame
}
\description{
Bind rows of multiple result tables into one long table
}
\examples{
results1 <- deseq2_to_dge(dds1)
results2 <- deseq2_to_dge(dds2)
merged <- merge_results_long(results1, results2,
                             result_names = c("Treatment1", "Treatment2"))

}
