% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results_wide.R
\name{merge_results_wide}
\alias{merge_results_wide}
\title{Merge multiple result tables into wide format}
\usage{
merge_results_wide(...)
}
\arguments{
\item{...}{List of result data frames/tibbles}
}
\value{
Merged wide-format data frame
}
\description{
Bind rows of multiple result tables and pivot to wide format
}
\examples{
results1 <- deseq2_to_dge(dds1)
results2 <- deseq2_to_dge(dds2)
merged <- merge_results_wide(results1, results2)

}
