% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_to_limma.R
\name{microarray_to_limma}
\alias{microarray_to_limma}
\title{Differential expression analysis for microarray data}
\usage{
microarray_to_limma(metadata, expression, design)
}
\arguments{
\item{metadata}{Metadata table}

\item{expression}{Microarray expression matrix}

\item{design}{Model design formula}
}
\value{
Fitted limma model object
}
\description{
Perform DE analysis on microarray data using limma
}
\examples{
metadata <- tibble(sample = c("s1","s2","s3"),
                   type = c("treated","untreated","treated"))
expression <- matrix(rnorm(9*200), nrow = 9, ncol = 200) 
design <- ~ type
limma_model <- microarray_to_limma(metadata, expression, design)

}
