% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA plot for RNAseq count data}
\usage{
pca_plot(metadata, counts, annotation = NULL)
}
\arguments{
\item{metadata}{A tibble with sample metadata}

\item{counts}{A tibble with RNAseq count data}

\item{annotation}{Column in metadata to use for annotation (default NULL)}
}
\value{
A ggplot2 PCA plot
}
\description{
Generates a PCA plot colored by a sample annotation
}
\examples{
metadata <- tibble(sample = c("s1", "s2", "s3"),
                   tissue = c("liver", "brain", "kidney"))  
counts <- tibble(gene_id = c("g1", "g2"),
                 s1 = c(10, 20), 
                 s2 = c(15, 50),
                 s3 = c(20, 100))
pca_plot(metadata, counts, annotation = "tissue")

}
