% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea_results.R
\name{plot_gsea_results}
\alias{plot_gsea_results}
\title{Visualize GSEA results}
\usage{
plot_gsea_results(
  gsea.res,
  top_n_pathways = 10,
  padj_cutoff = 0.1,
  remove_first_word_in_pathway_name = T
)
}
\arguments{
\item{gsea.res}{Data frame of GSEA results}

\item{top_n_pathways}{Number of up/downregulated pathways to include (default 10)}

\item{padj_cutoff}{Significance cutoff for shape (default 0.1)}

\item{remove_first_word_in_pathway_name}{Remove first word in pathway name (default TRUE)}
}
\value{
ggplot object
}
\description{
Generate a forest plot visualization for GSEA results
}
\examples{
results <- dge_to_gsea(de_results)
plot <- plot_gsea_results(results)

}
