% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_gsea_results.R
\name{plot_merged_gsea_results}
\alias{plot_merged_gsea_results}
\title{Visualize merged GSEA results}
\usage{
plot_merged_gsea_results(
  merged.gsea.results,
  padj_cutoff = 0.1,
  top_n_pathways = 15
)
}
\arguments{
\item{merged.gsea.results}{Data frame with merged GSEA results}

\item{padj_cutoff}{Significance cutoff for shape}

\item{top_n_pathways}{Number of top pathways per analysis to include}
}
\value{
ggplot object
}
\description{
Create a dot plot to compare GSEA results
}
\examples{
merged_results <- merge_gsea(results1, results2)
plot <- plot_merged_gsea_results(merged_results)

}
