% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generate volcano plot for differential expression results}
\usage{
plot_volcano(
  deseq.res,
  col.up = "red",
  col.down = "blue",
  gene.labels = NULL,
  n.labels = 10,
  fcCutoff = 0.5,
  fdrCutoff = 0.1,
  plot.title = ""
)
}
\arguments{
\item{deseq.res}{Tidy DE results table}

\item{col.up}{Color for upregulated points (default 'red')}

\item{col.down}{Color for downregulated points (default 'blue')}

\item{gene.labels}{Vector of gene labels to highlight (default NULL)}

\item{n.labels}{Number of top genes to label (default 10)}

\item{fcCutoff}{Fold change cutoff for significance (default 0.5)}

\item{fdrCutoff}{FDR cutoff for significance (default 0.1)}

\item{plot.title}{Plot title text (default '')}
}
\value{
ggplot2 object with volcano plot
}
\description{
Generate volcano plot for differential expression results
}
\examples{
results <- deseq2_to_dge(dds, contrast=c(1,0))
volcano_plot <- plot_volcano(results)

}
