% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_to_deseq2.R
\name{rnaseq_to_deseq2}
\alias{rnaseq_to_deseq2}
\title{Differential expression analysis with DESeq2}
\usage{
rnaseq_to_deseq2(metadata, counts, design)
}
\arguments{
\item{metadata}{Metadata table}

\item{counts}{Count data}

\item{design}{Model design formula}
}
\value{
DESeq2 result object
}
\description{
Perform differential expression analysis on count data
}
\examples{
metadata <- tibble(sample = c("s1","s2","s3"), 
                   type = c("treated","untreated","treated"))
counts <- tibble(gene_id = c("ENSG1","ENSG2"),
                 s1 = c(10,20),
                 s2 = c(15,50),
                 s3 = c(20,100)) 
design <- ~ type 
de_results <- rnaseq_to_deseq2(metadata, counts, design)

}
