% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_to_dream.R
\name{rnaseq_to_dream}
\alias{rnaseq_to_dream}
\title{Differential expression analysis with Dream}
\usage{
rnaseq_to_dream(
  metadata,
  counts,
  design,
  filterByExpr.design = ~1,
  ddf = "Kenward-Roger"
)
}
\arguments{
\item{metadata}{Metadata table}

\item{counts}{Count data table}

\item{design}{Model design formula}

\item{ddf}{Method to estimate degrees of freedom}
}
\value{
Limma model fit object
}
\description{
Perform differential expression analysis with mixed effects models on RNA-seq count data using Dream
}
\examples{
metadata <- readRDS("metadata.rds")
counts <- readRDS("counts.rds")
design <- ~ condition + batch
model <- rnaseq_to_dream(metadata, counts, design)

}
