% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsva.R
\name{run_gsva}
\alias{run_gsva}
\title{Perform gene set variance analysis}
\usage{
run_gsva(counts, metadata, design, gsets)
}
\arguments{
\item{counts}{Count data as tibble}

\item{metadata}{Sample metadata as tibble}

\item{design}{Model design formula}

\item{gsets}{Gene set database as list of gene vectors}
}
\value{
Tibble with sample scores for each gene set
}
\description{
Run GSVA on count data to assess pathway activity
}
\examples{
# Load example data
counts <- get_example_counts() 
metadata <- get_example_metadata()
design <- ~ condition
gsets <- get_example_genesets()

# Run GSVA
gsva_results <- run_gsva(counts, metadata, design, gsets)

}
