% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate metadata and gene expression data}
\usage{
validate_data(
  metadata,
  gene_expression,
  sample_name = "sampleid",
  gene_id = "ensembl_gene_id"
)
}
\arguments{
\item{metadata}{Tibble with sample metadata}

\item{gene_expression}{Tibble with gene expression data}

\item{sample_name}{Name of sample ID column in metadata (default 'sampleid')}

\item{gene_id}{Name of gene ID column in expression (default 'ensembl_gene_id')}
}
\value{
TRUE if validation passed, FALSE otherwise
}
\description{
Confirm metadata and gene expression data formats are compatible for analysis
}
\examples{
metadata <- tibble(sampleid = c("s1", "s2", "s3"))
expression <- tibble(ensembl_gene_id = c("ENSG1", "ENSG2"),
                      s1 = c(1,2),
                      s2 = c(2,1), 
                      s3 = c(3,5))
validate_data(metadata, expression) # Returns TRUE

}
