% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_to_apeglm_dge.R
\name{deseq2_to_apeglm_dge}
\alias{deseq2_to_apeglm_dge}
\title{Extract shrunken DESeq2 results (apeglm)}
\usage{
deseq2_to_apeglm_dge(
  dds,
  ihw.fdrs = TRUE,
  var_interest,
  numerator = NULL,
  denominator = NULL,
  use_adaptiveshrink = TRUE,
  save_output = NULL,
  output_filename = NULL
)
}
\arguments{
\item{dds}{DESeqDataSet after DE analysis}

\item{ihw.fdrs}{Use the IHW package to calculate FDRs (default TRUE)}

\item{var_interest}{Variable of interest}

\item{numerator}{Numerator value (default NULL)}

\item{denominator}{Denominator value (default NULL)}

\item{use_adaptiveshrink}{Logical for adaptive prior (default TRUE)}

\item{save_output}{Logical to save results (default NULL)}

\item{output_filename}{Custom filename for results (default NULL)}
}
\value{
Shrunken DESeq2 results as tidy table
}
\description{
Extract shrunken DESeq2 results (apeglm)
}
\examples{
\dontrun{
options(save.output = TRUE)
results <- deseq2_to_apeglm_dge(dds, save_output = FALSE)  
}

}
