% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_to_dge.R
\name{limma_to_dge}
\alias{limma_to_dge}
\title{Convert limma results to tidy table}
\usage{
limma_to_dge(model, contrast)
}
\arguments{
\item{model}{Fitted limma model object}

\item{contrast}{Contrast to extract results for}
}
\value{
Tidy tibble with DE results
}
\description{
Convert limma results to tidy table
}
\examples{
metadata <- tibble(sample = c("s1","s2","s3"),
                   type = c("treated","untreated","treated"))  
expression <- matrix(rnorm(9*200), nrow = 9, ncol = 200)
design <- model.matrix(~ type, data = metadata) 
limma_model <- lmFit(expression, design)
limma_model <- eBayes(limma_model)
 
# Get results for treated vs untreated
de_table <- limma_to_dge(limma_model, contrast = c(-1,1,0)) 

}
