% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merged_gsea_results.R
\name{plot_merged_gsea_results}
\alias{plot_merged_gsea_results}
\title{Visualize merged GSEA results}
\usage{
plot_merged_gsea_results(
  merged_long_gsea_results,
  padj_cutoff = 0.1,
  top_n_pathways = 15,
  minimum_significant_groups = 1,
  str_wrap_legnth = 40,
  str_trunc_length = 1000
)
}
\arguments{
\item{padj_cutoff}{Significance FDR cutoff (default 0.1)}

\item{top_n_pathways}{Number of top pathways per analysis to include (default 15)}

\item{minimum_significant_groups}{Number of groups with a significant result for included pathways (default 1 )}

\item{merged.gsea.results}{Data frame with merged GSEA results}
}
\value{
ggplot object
}
\description{
Create a dot plot to compare GSEA results
}
\examples{
merged_results <- merge_gsea(results1, results2)
plot <- plot_merged_gsea_results(merged_results)

}
