% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generate volcano plot for differential expression results}
\usage{
plot_volcano(
  dge.res,
  col_up = "red",
  col_down = "blue",
  gene_label_col = "hgnc_symbol",
  gene_labels = NULL,
  n_labels = 10,
  fc_cutoff = 0.5,
  fdr_cutoff = 0.1,
  plot_title = ""
)
}
\arguments{
\item{dge.res}{Tidy DE results table}

\item{col_up}{Color for upregulated points (default 'red')}

\item{col_down}{Color for downregulated points (default 'blue')}

\item{gene_label_col}{Column with gene labels to highlight (default NULL)}

\item{gene_labels}{Vector of gene labels to highlight (default NULL)}

\item{n_labels}{Number of top genes to label (default 10)}

\item{fc_cutoff}{Fold change cutoff for significance (default 0.5)}

\item{fdr_cutoff}{FDR cutoff for significance (default 0.1)}

\item{plot_title}{Plot title text (default '')}
}
\value{
ggplot2 object with volcano plot
}
\description{
Generate volcano plot for differential expression results
}
\examples{
results <- deseq2_to_dge(dds, contrast=c(1,0))
volcano_plot <- plot_volcano(results)

}
