% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_to_limma.R
\name{rnaseq_to_limma}
\alias{rnaseq_to_limma}
\title{Differential expression analysis with limma-voom}
\usage{
rnaseq_to_limma(metadata, counts, design, filterByExpr.design = ~1)
}
\arguments{
\item{metadata}{Metadata table}

\item{counts}{Count data table}

\item{design}{Model design formula}
}
\value{
Limma model fit object
}
\description{
Perform differential expression analysis on RNA-seq count data using limma-voom
}
\examples{
metadata <- readRDS("metadata.rds")
counts <- readRDS("counts.rds")
design <- ~ condition + batch
model <- rnaseq_to_limma(metadata, counts, design)

}
