% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate metadata and expression data}
\usage{
validate_data(metadata, expression)
}
\arguments{
\item{metadata}{Metadata table}

\item{expression}{Gene expression data}
}
\value{
NULL if validation passes, stops with error otherwise
}
\description{
Check that metadata and expression formats are compatible
}
\examples{
metadata <- readRDS("metadata.rds")
expression <- readRDS("expression.rds")
validate_data(metadata, expression)

}
